🐛 Fix client import of resolvePreviewImagePresence
Move the helper to a client-safe module so the upload route component no longer pulls fetchPreviews.server into the browser bundle.
